Fix stupid bug in finder hack.
We need to distinguish between the source loader and other loaders;
the obvious quick&dirty way to do this doesn't work (because you
usually have a frozen version of SourceFileLoader, which doesn't
match the one you can import), so we get even dirtier by checking
its type name.

As with everything in this project, don't do this in real life.